% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtapi.R
\name{checkinputargs}
\alias{checkinputargs}
\title{Jterator API.
Checks input arguments for correct class and attributes.}
\usage{
checkinputargs(input_args)
}
\arguments{
\item{input_args}{List of input arguments.}
}
\value{
List of checked input arguments.
}
\description{
Jterator API.
Checks input arguments for correct class and attributes.
}

