% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtapi.R
\name{gethandles}
\alias{gethandles}
\title{Jterator API.
Reading "handles" stream from YAML file.}
\usage{
gethandles(handles_stream)
}
\arguments{
\item{handles_stream}{YAML stream from standard input.}
}
\value{
List of module input/output information.
}
\description{
Jterator API.
Reading "handles" stream from YAML file.
}

