% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtapi.R
\name{readconfig}
\alias{readconfig}
\title{Jterator API.
Reading configuration settings.}
\usage{
readconfig(configuration)
}
\arguments{
\item{configuration}{YAML configuration settings}
}
\value{
list
}
\description{
Jterator API.
Reading configuration settings.
}

