% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtapi.R
\name{readinputargs}
\alias{readinputargs}
\title{Jterator API.
Reading input arguments from HDF5 file
using the locations specified in "handles".}
\usage{
readinputargs(handles)
}
\arguments{
\item{handles}{List of module input/output information.}
}
\value{
List of input arguments read from HDF5 file.
}
\description{
Jterator API.
Reading input arguments from HDF5 file
using the locations specified in "handles".
}

