% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jtapi.R
\name{writedata}
\alias{writedata}
\title{Jterator API.
Writing data to HDF5 file using the location specified in "handles".}
\usage{
writedata(data, data_file)
}
\arguments{
\item{data}{List of data output that should be written to HDF5 file.}

\item{data_file}{List of module input/output information.}
}
\description{
Jterator API.
Writing data to HDF5 file using the location specified in "handles".
}

